+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Added: support for `@media` queries
+
 # 5.1.0 - 2016-01-07
 
 - Added: "warnWhenCannotResolve" option to warn when calc() are not reduced to a single value